Edited some code and added a working chunk manager and chunk loader, which is running in a separated thread for better performance.

+++ src/de/meinkraft/Chunk.java
@@ -5,8 +5,6 @@ import static org.lwjgl.opengl.GL11.GL_TRIANGLES;
 import java.nio.FloatBuffer;
 import java.nio.IntBuffer;
 
-import org.lwjgl.BufferUtils;
-
 import de.meinkraft.lib.VAO;
 
 public class Chunk {
@@ -15,7 +13,11 @@ public class Chunk {
 	
 	private final Block[][][] blocks;
 	private VAO vao;
-	private int count;
+	public FloatBuffer vertices;
+	public IntBuffer indices;
+	public int solidBlocks;
+	
+	private boolean doLoad;
 	
 	private final int x, z;
 	
@@ -24,92 +26,36 @@ public class Chunk {
 		this.z = z;
 		
 		blocks = new Block[SIZE_X][SIZE_Y][SIZE_Z];
-		vao = new VAO(GL_TRIANGLES);
+	}
+	
+	public void update() {
+		if(doLoad)
+			load();
 	}
 	
 	public void render() {
 		if(isLoaded())
-			vao.render(count, 0);
+			vao.render(solidBlocks, 0);
 	}
 	
 	public void load() {
 		if(vao == null)
 			vao = new VAO(GL_TRIANGLES);
 		
-		int verticesCount = 0;
-		for(int x = 0; x < Chunk.SIZE_X; x++)
-			for(int y = 0; y < Chunk.SIZE_Y; y++)
-				for(int z = 0; z < Chunk.SIZE_Z; z++) {
-					Blockcheck bc = new Blockcheck();
-					
-					if(getBlockAt(x + 1, y, z).getType().isSolid() && getBlockAt(x + 1, y, z).getMaterial().isOpaque())
-						bc.east = true;
-					
-					if(getBlockAt(x - 1, y, z).getType().isSolid() && getBlockAt(x - 1, y, z).getMaterial().isOpaque())
-						bc.west = true;
-					
-					if(getBlockAt(x, y + 1, z).getType().isSolid() && getBlockAt(x, y + 1, z).getMaterial().isOpaque())
-						bc.top = true;
-					
-					if(getBlockAt(x, y - 1, z).getType().isSolid() && getBlockAt(x, y - 1, z).getMaterial().isOpaque())
-						bc.bottom = true;
-					
-					if(getBlockAt(x, y, z + 1).getType().isSolid() && getBlockAt(x, y, z + 1).getMaterial().isOpaque())
-						bc.south = true;
-					
-					if(getBlockAt(x, y, z - 1).getType().isSolid() && getBlockAt(x, y, z - 1).getMaterial().isOpaque())
-						bc.north = true;
-					
-					verticesCount += getBlockAt(x, y, z).getType().getVerticesCount(bc);
-				}
-		
-		FloatBuffer vertices = BufferUtils.createFloatBuffer(verticesCount * 6);
-		for(int x = 0; x < Chunk.SIZE_X; x++)
-			for(int y = 0; y < Chunk.SIZE_Y; y++)
-				for(int z = 0; z < Chunk.SIZE_Z; z++) {
-					Blockcheck bc = new Blockcheck();
-					
-					if(getBlockAt(x + 1, y, z).getType().isSolid() && getBlockAt(x + 1, y, z).getMaterial().isOpaque())
-						bc.east = true;
-					
-					if(getBlockAt(x - 1, y, z).getType().isSolid() && getBlockAt(x - 1, y, z).getMaterial().isOpaque())
-						bc.west = true;
-					
-					if(getBlockAt(x, y + 1, z).getType().isSolid() && getBlockAt(x, y + 1, z).getMaterial().isOpaque())
-						bc.top = true;
-					
-					if(getBlockAt(x, y - 1, z).getType().isSolid() && getBlockAt(x, y - 1, z).getMaterial().isOpaque())
-						bc.bottom = true;
-					
-					if(getBlockAt(x, y, z + 1).getType().isSolid() && getBlockAt(x, y, z + 1).getMaterial().isOpaque())
-						bc.south = true;
-					
-					if(getBlockAt(x, y, z - 1).getType().isSolid() && getBlockAt(x, y, z - 1).getMaterial().isOpaque())
-						bc.north = true;
-					
-					getBlockAt(x, y, z).getType().addVertices(getX() * Chunk.SIZE_X + x, y, getZ() * Chunk.SIZE_Z + z, vertices, bc, 1,1,1,1,1,1);
-				}
-		vertices.flip();
-		
-		IntBuffer indices = BufferUtils.createIntBuffer(vertices.limit() / 4 * 6);
-		for(int i = 0; i < indices.limit(); i += 6) {
-			indices.put(i + 0);
-			indices.put(i + 1);
-			indices.put(i + 2);
-			indices.put(i + 0);
-			indices.put(i + 2);
-			indices.put(i + 3);
-		}
-		indices.flip();
-		
-		count = indices.limit();
 		vao.initVBO(vertices);
 		vao.initIBO(indices);
 		vao.initVAO("3 6 0", "3 6 3");
+		
+		solidBlocks = indices.limit();
+		vertices = null;
+		indices = null;
+		doLoad = false;
 	}
 	
 	public void unload() {
-		vao.delete();
+		if(isLoaded())
+			vao.delete();
+		
 		vao = null;
 	}
 	
@@ -120,7 +66,7 @@ public class Chunk {
 	public boolean isLoaded() {
 		return vao != null;
 	}
-	
+
 	public Block getBlockAt(int x, int y, int z) {
 		if(x >= 0 && x < SIZE_X && y >= 0 && y < SIZE_Y && z >= 0 && z < SIZE_Z)
 			return blocks[x][y][z];
@@ -133,6 +79,10 @@ public class Chunk {
 			blocks[x][y][z] = block;
 	}
 	
+	public void doLoad() {
+		doLoad = true;
+	}
+
 	public int getX() {
 		return x;
 	}
